+++ WPF.Net.Examples/Views/PageHttpClient.xaml.cs
@@ -8,7 +8,7 @@ namespace WPF.Net.Examples.Views
     {
         #region Private fields and properties
 
-        private AppSettings _appSet;
+        private AppSettings AppSettings { get; set; }
 
         #endregion
 
@@ -25,14 +25,14 @@ namespace WPF.Net.Examples.Views
 
         private void PageHttpService_Loaded(object sender, RoutedEventArgs e)
         {
-            _appSet = ViewModels.Utils.GetSettings(this);
+            AppSettings = ViewModels.Utils.GetSettings(this);
         }
 
         private void ButtonHttpGet_OnClick(object sender, RoutedEventArgs e)
         {
             var task = Task.Run(async () =>
             {
-                await _appSet.HttpClient.OpenAsync(_appSet.Proxy).ConfigureAwait(false);
+                await AppSettings.HttpClient.OpenAsync(AppSettings.Proxy).ConfigureAwait(false);
             });
             task.Wait();
         }